+++ client/src/components/comment/CommentList.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Comment from "./Comment";
+import Comment from "../comment/Comment";
 import { IComment, ICommentPage } from "../../models/comment.model";
 import { InfiniteData } from "@tanstack/react-query";
