+++ client/src/components/community/CommunityComments.tsx
@@ -1,8 +1,8 @@
 import "../../styles/css/components/community/comments.css";
 import useCommunityComment from "../../hooks/useCommunityComment";
 import { useIntersectionObserver } from "../../hooks/useIntersectionObserver";
-import CommentList from "./CommentList";
-import CommentsEmpty from "./CommentsEmpty";
+import CommentList from "../comment/CommentList";
+import CommentsEmpty from "../comment/CommentsEmpty";
 
 interface IProps {
   postId: number;
