+++ client/src/pages/event/EventDetail.tsx
@@ -2,8 +2,8 @@ import { Suspense } from "react";
 import "../../styles/css/pages/event/eventDetail.css";
 import { useParams } from "react-router-dom";
 import ErrorNotFound from "../../components/error/ErrorNotFound";
-import PostDetail from "../../components/community/PostDetail";
-import CommentForm from "../../components/community/CommentForm";
+import PostDetail from "../../components/communityAndEvent/PostDetail";
+import CommentForm from "../../components/comment/CommentForm";
 import useEvent from "../../hooks/useEvent";
 import useEventComment from "../../hooks/useEventComment";
 import EventComments from "../../components/event/EventComments";
