+++ client/src/styles/css/components/communityAndEvent/postDetail.css
