+++ client/src/styles/scss/components/common/avartar.scss
