+++ client/src/styles/scss/components/communityAndEvent/post.scss
