+++ client/src/styles/scss/components/communityAndEvent/postDetail.scss
