+++ client/src/styles/scss/components/communityAndEvent/postEmpty.scss
